// packages/main/src/types/api/ticket.ts
export interface Lead {
    id: number;
    text_id: string;
    form_id: null;
    first_name: string;
    last_name: string;
    email: string;
    gender: string;
    phone: string;
    created_at: Date;
    updated_at: Date;
    interest: null;
    smoker: boolean;
    medical_condition: null;
    age: null;
    height_feet: null;
    height_inch: null;
    weight: null;
    stage: string;
    date_of_birth: null;
    source: string;
    contacted: boolean;
    project_id: number;
    address1: null;
    address2: null;
    city: null;
    state: null;
    postal_code: null;
    country: null;
    financing_link_clicked: null;
    financing_company: null;
    message: null;
    staff_id: null;
    location_id: null;
    review: null;
    email_optin: number;
    session_ended: boolean;
    bmi: null;
    messages?: Message[];
    staff?: Staff;
    notes?: Note[];
    isChecked?: boolean;
    email_optout: boolean;
    sms_optout: boolean;
}
// ...
export enum MessageableTypeEnum {
    AppModelsEmailMessage = "App\\Models\\EmailMessage",
    AppModelsWidgetMessage = "App\\Models\\WidgetMessage",
    AppModelsFileMessage = "App\\Models\\FileMessage",
}
export enum MessageType {
    Inbound = "inbound",
    Outbound = "outbound",
}

export interface Message {
    id: number;
    type: MessageType;
    metadata: null;
    messageable_type: MessageableTypeEnum;
    messageable_id: number;
    created_at: Date;
    updated_at: Date;
    channel_id: number;
    staff_id: number | null;
    message: null | string;
    to: To[] | null;
    from_system: boolean;
    pivot: LeadPivot;
    messageable: Messageable;
    staff: Staff | null;
    channel: Channel;
    recipients: Lead[];
    // eslint-disable-next-line @typescript-eslint/no-explicit-any
    cc: any[];
    is_conversation: boolean;
}

export interface LeadPivot {
    lead_id: number;
    message_id: number;
}
export interface To {
    lead_id?: number;
    email?: string;
}

export interface Messageable {
    id: number;
    created_at: Date;
    updated_at: Date;
    question: null | string;
    appointment_id: number | null;
    type: null;
    estimate_id: number | null;
    qualification_id: number | null;
    qualification?: Qualification;
    appointment?: Appointment;
    estimate?: Estimate;
    files?: File[];
    plain?: string;
    html?: string;
    clean_html?: string;
    from?: null;
    cc?: To[];
    subject?: string;
    thread_id?: number;
    file?: File;
    file_id?: number;
}
// ...
export interface Note {
    note: string;
    noteable_id: number;
    noteable_type: string;
    noteable?: Lead;
    updated_at: Date;
    created_at: Date;
    id: number;
    ticket_id: number;
    staff_id: number;
    staff: Staff;
    mentions: Mention[];
}

export interface Mention {
    id: number;
    note_id: number;
    seen: boolean;
    note?: Note;
    mentionable?: Staff;
    updated_at: Date;
    created_at: Date;
}